Stop Using type_traits details.
Some of the type_traits internal details this header depends on will either go away or be moved.  Either way, best not to rely on another libraries internal details.  In the short term the old code will work but will generate a ton of warnings about using a deprecated header.